Delete pegawai

+++ app/Http/Controllers/PegawaiController.php
@@ -49,19 +49,16 @@ class PegawaiController extends Controller
         $pegawai->tanggal_lahir = Carbon::createFromFormat('m/d/Y', $request->get('tanggal_lahir'))->format('Y-m-d');
         $pegawai->jk = $request->get('jk');
         $pegawai->agama = $request->get('agama');
-        $pegawai->phone = Str::replaceFirst('0','+62',$request->get('phone')) ;
+        $pegawai->phone = $request->get('phone') ;
         $pegawai->alamat = $request->get('alamat');
 
         if ( $request->file('foto') ) {
             $foto = $request->file('foto')->store('fotos', 'public');
-
-            $pegawai->foto = 'fotos/'. $request->file('foto')->getClientOriginalName();
-            return response()->json($pegawai->foto);
-
+            $pegawai->foto = $foto;
         }
 
         $pegawai->save();
-        return redirect()->route('pegawais.create')->with('status', 'Data Has Been Saved Successfully.');
+        return redirect()->route('pegawais.create')->with('status', 'Data has been saved successfully.');
     }
 
     /**
@@ -72,8 +69,6 @@ class PegawaiController extends Controller
      */
     public function show(Pegawai $pegawai)
     {
-        // return response()->json($pegawai);
-
         return view('pegawais.show', compact('pegawai'));
     }
 
@@ -95,7 +90,7 @@ class PegawaiController extends Controller
      * @param  \App\Pegawai  $pegawai
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Pegawai $pegawai)
+    public function update(PegawaiRequest $request, Pegawai $pegawai)
     {
 
         $pegawai->nama_lengkap = $request->get('nama_lengkap');
@@ -103,19 +98,19 @@ class PegawaiController extends Controller
         $pegawai->tanggal_lahir = Carbon::createFromFormat('m/d/Y', $request->get('tanggal_lahir'))->format('Y-m-d');
         $pegawai->jk = $request->get('jk');
         $pegawai->agama = $request->get('agama');
-        $pegawai->phone = Str::replaceFirst('0','+62',$request->get('phone')) ;
+        $pegawai->phone = $request->get('phone');
         $pegawai->alamat = $request->get('alamat');
-        return response()->json($request->foto);
 
-        if ( $pegawai->foto && file_exists(storage_path('app/public/' . $pegawai->foto ))) {
-            Storage::delete('public/'. $pegawai->foto);
+        if ( $request->file('foto')) {
+            if ( $pegawai->foto && file_exists(storage_path('app/public/' . $pegawai->foto ))) {
+                Storage::delete('public/'. $pegawai->foto);
+            }
             $foto = $request->file('foto')->store('fotos', 'public');
             $pegawai->foto = $foto;
-            return response()->json($pegawai);
         }
 
         $pegawai->save();
-        return redirect()->route('pegawais.index')->with('status', 'Data Has Been Updated Successfully');
+        return redirect()->route('pegawais.index')->with('status', 'Data has been updated successfully');
     }
 
     /**
@@ -126,6 +121,7 @@ class PegawaiController extends Controller
      */
     public function destroy(Pegawai $pegawai)
     {
-        //
+        $pegawai->delete();
+        return back()->with('success', 'Data has been removed');
     }
 }
